refactor: deduplicate parser utilities across imageParser, textBoxParser, shapeParser
Extract shared DrawingML parsing functions into drawingUtils.ts and consolidate
duplicated utilities into their canonical locations:

- emuToPixels: 3 copies → 1 in units.ts (now null-safe with Math.round)
- findByFullName: 3 copies → 1 in xmlParser.ts
- findAllByLocalName: 2 copies → use existing findChildrenByLocalName
- parsePositionH/V/AnchorPosition: 3 copies → 1 in drawingUtils.ts
- parseWrap/parseWrapElement: 3 copies → 1 in drawingUtils.ts
- parseColorElement + applyColorModifiers: 2 copies → 1 in drawingUtils.ts
- parseFill/parseOutline: 2 copies → 1 in drawingUtils.ts
- themeColorMap: 4 inline copies → resolveColorValueToHex in drawingUtils.ts
- resolveFillColor/resolveOutlineColor: simplified to use shared resolver

Also:
- Add early-exit guard in enrichParagraphTextBoxes (skip empty paragraphs)
- Extract DEFAULT_TEXTBOX_WIDTH constant (was magic number 200 in 3 places)

Net result: -972 lines of duplicated code removed.
