+++ packages/core/src/layout-engine/types.ts
@@ -401,6 +401,9 @@ export type ColumnBreakBlock = {
 /** Default internal margins for text boxes (OOXML defaults in pixels) */
 export const DEFAULT_TEXTBOX_MARGINS = { top: 4, bottom: 4, left: 7, right: 7 };
 
+/** Default text box width in pixels when no width is specified */
+export const DEFAULT_TEXTBOX_WIDTH = 200;
+
 /**
  * Text box block — positioned container with paragraph content.
  */
